local_web_portal/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
jinja2==3.1.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==45.0.7
python-dotenv==1.1.1
itsdangerous==2.2.0
